Refactor adapt functionality
